refactor: 이름 검증 로직 Players에 위임
