+++ src/main/java/blackjack/domain/Blackjack.java
@@ -7,28 +7,14 @@ import blackjack.domain.dto.ResultDto;
 import blackjack.domain.strategy.NumberGenerator;
 
 public class Blackjack {
-	private static final String DUPLICATED_ERROR = "[ERROR] 이름은 중복될 수 없습니다.";
-	private static final String NO_PLAYER_ERROR = "[ERROR] 플레이어는 1명 이상이여야 합니다.";
 	private static final int NUMBER_OF_INIT_CARD = 2;
 
 	private final Players players;
 	private final Dealer dealer;
 
 	public Blackjack(List<String> playerNames) {
-		validateNames(playerNames);
 		this.players = new Players(playerNames);
 		this.dealer = new Dealer();
-	}
-
-	private void validateNames(List<String> playerNames) {
-		if (playerNames.size() == 0) {
-			throw new IllegalArgumentException(NO_PLAYER_ERROR);
-		}
-
-		List<String> duplicatedChecker = playerNames.stream().distinct().collect(Collectors.toList());
-		if(duplicatedChecker.size() != playerNames.size()) {
-			throw new IllegalArgumentException(DUPLICATED_ERROR);
-		}
 	}
 
 	public void distributeInitialCards(NumberGenerator numberGenerator) {
